Fix errors that occur when IDN is installed
IDN throws a exception that claims 'Output is too small or too
large' when a segment of the host is longer than 63 characters.

In order to test IDN correctly it must be included as a development
dependency, but since IDN is not required to use Addressable, it is
commented out as a development dependency. In order to fully test
the gem the specs must be run with and without IDN.